docs - Bind API to all the interfaces of the container
Fixes the issues that requests could not be sent from the host in Windows and Mac by binding the API to all the container interfaces instead of only localhost.
This may create issues with get_all_appointments, since it can only be called from inside the container. Not a big deal since that's more of a dev endpoint, but it'll
be nice to give it a fix eventually.