test - fix e2e*
There's an issue with the current version of the e2e tests not killingthe gunicorn process when the main process is being killed. I think this can be solved by shuting down the daemon with the rpc command instead of a terminate, but that has to be done in a follow-up.