teos: Updates how the API process is created
Python 3.8 processes are created using spawn for MacOS instead of fork (which use to be the default). The previous version of the code was passing some data that cannot be serialised (pickled) when spawn is used, so it needed to be reworked.

Forcing fork instead of spawn may have also worked, but Windows does not support it, so the old version of the code would have failed on Windows anyway.