lnnet - Changes defaults from None
Several optional values were defaulted to None, this mean that checking some properties of this values required checking if they were None first (e.g. checking the length of the  for a ). This sets the default value to the same type as expected for non-defaults, so the type can be assumed in every case