teos: Handles subscription not being properly expired
Subscriptions were only expired if the slots were full, but not if the expiry time was reached.

The user info handling and cleanup has been moved to the Gatekeeper, that has been hooked to the ChainMonitor. In this way the modules are more independent, and the Watcher and Responder do not need to fight to delete the expired users from the Gatekeeper.