teos: Adds bitcoind polling to ensure it is alive
The ChainMonitor polling thread now is also used to make sure bitcoind has not crashed. Locks have been added so the ChainMonitor thread can lock other background threads if bitcoind is down. This does not apply to user requests, since they need to give a timely response to the user.