plugin: Adds c-lightning options to plugin
The plugin can now be configured with c-lightning options, meaning that the config parameters can be set on the c-lightning config file so no standalone config file is required.

For now, the standalone config file is left, so both options are available. The current priority is: c-lightning conf, standalone conf, defaults.