teos - Fixes logs not showing if teos shutdowns right after start
In some OSs teos was not showing logs if the system was not properly configured, due to it shutting down rather quickly. The root of the issue if the logging server shutting down before it has the chance of logging the info.

This commit syncs the shutting down process in case of an exception so the logging server can output properly.

Also simplifies the logging server (removes unnecessary methods, calls super properly, ...)